[NFC] IRGen: Move these functions.
